Starting to factor out attribute accumulation into its own file
